requirements.txt
streamlit
fpdf2
plotly
pandas
